Translate PessoaSpecs
