+++ TestePratico.Domain/Entities/Specifications/PersonSpecs/PersonCantHaveBirthdateInTheFutureSpec.cs
@@ -1,13 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using TestePratico.Domain.Interfaces.Validation;
 
-namespace TestePratico.Domain.Entities.Specifications.PessoaSpecs
+namespace TestePratico.Domain.Entities.Specifications.PersonSpecs
 {
-    public class PessoaNaoDeveTerDataDeNascimentoNoFuturoSpec : ISpecification<Person>
+    public class PersonCantHaveBirthdateInTheFutureSpec : ISpecification<Person>
     {
         public string Field => nameof(Person.Birthdate);
