--- TestePratico.Domain/Entities/Specifications/PessoaSpecs/PessoaPrecisaTerUmaUFDefinidaSpec.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TestePratico.Domain.Interfaces.Validation;
-
-namespace TestePratico.Domain.Entities.Specifications.PessoaSpecs
-{
-    public class PessoaPrecisaTerUmaUFDefinidaSpec : ISpecification<Person>
-    {
-        public string Field => nameof(Person.UFId);
-
-        public bool IsSatisfiedBy(Person entity)
-        {
-            return entity.UFId.HasValue;
-        }
-    }
-}
