+++ TestePratico.Domain/Entities/Validation/PersonValidator.cs
@@ -1,4 +1,4 @@
-﻿using TestePratico.Domain.Entities.Specifications.PessoaSpecs;
+﻿using TestePratico.Domain.Entities.Specifications.PersonSpecs;
 using TestePratico.Domain.Validation;
 
 namespace TestePratico.Domain.Entities.Validation
@@ -7,10 +7,10 @@ namespace TestePratico.Domain.Entities.Validation
     {
         public PersonValidator()
         {
-            AddRule(new ValidationRule<Person>(new PessoaPrecisaTerUmNomeDefinidoSpec(), "It is required to provide a name"));
-            AddRule(new ValidationRule<Person>(new PessoaPrecisaTerUmNomeEntre2e150Caracteres(), "Name must have between 2 and 150 characters"));
-            AddRule(new ValidationRule<Person>(new PessoaPrecisaTerUmaUFDefinidaSpec(), "It's required to select an UF"));
-            AddRule(new ValidationRule<Person>(new PessoaNaoDeveTerDataDeNascimentoNoFuturoSpec(), "Birthdate can't be in the future"));
+            AddRule(new ValidationRule<Person>(new PersonMustHaveANameDefinedSpec(), "It is required to provide a name"));
+            AddRule(new ValidationRule<Person>(new PersonMustHaveANameBetween2and150ChachactersLongSpec(), "Name must have between 2 and 150 characters"));
+            AddRule(new ValidationRule<Person>(new PersonMustHaveAUfDefinedSpec(), "It's required to select an UF"));
+            AddRule(new ValidationRule<Person>(new PersonCantHaveBirthdateInTheFutureSpec(), "Birthdate can't be in the future"));
         }
     }
 }
